Cate & Nelson Quadrants Analysis, 1971

This function runs the quadrants analysis suggested by Cate and Nelson (1965)

Usage

cate_nelson_1971(data = NULL, stv, ry, tidy = FALSE, plot = FALSE)

Arguments

data

argument to call a data.frame or data.table containing the data

stv

argument to call the vector or column containing the soil test value (stv) data

ry

argument to call the vector or column containing the relative yield (ry) data

tidy

logical operator (TRUE/FALSE) to decide the type of return. TRUE returns a data.frame, FALSE returns a list (default).

plot

logical operator (TRUE/FALSE) to decide the type of return. TRUE returns a ggplot, FALSE returns either a list (tidy == FALSE) or a data.frame (tidy == TRUE).

Value

it returns an object of type ggplot when plot == TRUE. If plot == FALSE, it returns an object of type data.frame if tidy == TRUE, otherwise, it returns a list it returns an object of type list containing the main results plus a ggplot object with the figure display

Details

See Cate, R. B. Jr., and L. A. Nelson. 1965. A rapid method for correlation of soil test analysis with plant response data. North Carolina Agric. Exp. Stn., International soil Testing Series l. No. 1. This code was adapted from Mangiafico, S. S. (2013). Cate-Nelson Analysis for Bivariate Data Using R-project. The Journal of Extension, 51(5), Article 33. https://tigerprints.clemson.edu/joe/vol51/iss5/33

See also

Examples

if (FALSE) {
if(interactive()){
 #EXAMPLE1
 }
}