Modified Arcsine-Log Calibration Curve

This function runs the modified arcsine-log calibration curve to estimate critical soil test values (CSTV) following Correndo et al. (2017)

Usage

mod_alcc(
  data = NULL,
  ry,
  stv,
  target,
  confidence = 0.95,
  tidy = FALSE,
  plot = FALSE
)

boot_mod_alcc(data, ry, stv, n = 500, target = 90, confidence = 0.95, ...)

Arguments

data

Optional argument to call and object of type data.frame or data.table containing the stv and ry data, Default: NULL

ry

name of the vector containing relative yield values (%) of type numeric.

stv

name of the vector containing soil test values of type numeric.

target

numeric value of relative yield target (e.g. 90 for 90%) to estimate the CSTV.

confidence

numeric value of confidence level (e.g. 0.95 for significance = 0.05)

tidy

logical operator (TRUE/FALSE) to decide the type of return. TRUE returns a data.frame, FALSE returns a list (default).

plot

logical operator (TRUE/FALSE) to decide the type of return. TRUE returns a ggplot, FALSE returns either a list (tidy == FALSE) or a data.frame (tidy == TRUE).

n

sample size for the bootstrapping Default: 500

...

when running bootstrapped samples, open arguments serve to add grouping Variables (factor or character) Default: NULL

Value

returns an object of type ggplot if plot = TRUE.

returns an object of class data.frame if tidy = TRUE,

returns an object of class list if tidy = FALSE.

boot_mod_alcc: bootstrapping function

Details

See online-documentation for additional details.

Note

For extended reference, we recommend to visit doi:10.7910/DVN/NABA57 and https://github.com/adriancorrendo/modified-ALCC by Adrian Correndo.

References

Correndo et al. (2017). A modification of the arcsine–log calibration curve for analysing soil test value–relative yield relationships. Crop and Pasture Science, 68(3), 297-304. doi:10.1071/CP16444

See also

Examples

# \donttest{
 # Example 1 dataset
 dat <- data.frame("ry" = c(65,80,85,88,90,94,93,96,97,95,98,100,99,99,100),
                   "stv" = c(1,2,3,4,5,6,7,8,9,10,11,12,13,14,15))
 # Run
 fit_example <- mod_alcc(data = dat, ry = ry, stv = stv, target=90, confidence = 0.95)
#> Warning: 7 STV points exceeded two-times (2x) 
#>   the CSTV for 90% of RY. Risk of leverage. You may consider a sensitivity analysis by 
#>   removing extreme points, re-run the mod_alcc(), and check results.
 fit_example
#> $n
#> [1] 15
#> 
#> $r
#> [1] 0.9682908
#> 
#> $target
#> [1] 90
#> 
#> $CSTV
#> [1] 4.478476
#> 
#> $LL
#> [1] 3.947041
#> 
#> $UL
#> [1] 5.081463
#> 
#> $confidence
#> [1] 0.95
#> 
#> $p_value
#> [1] 3.296044e-09
#> 
#> $CSTV90
#> [1] 4.478476
#> 
#> $n.90x2
#> [1] 7
#> 
#> $CSTV100
#> [1] 19.15054
#> 
#> $n.100
#> [1] 0
#> 
#> $Curve
#> $Curve[[1]]
#> # A tibble: 176 × 2
#>    RY.fitted STV.fitted
#>        <dbl>      <dbl>
#>  1      65         1.10
#>  2      65.2       1.11
#>  3      65.4       1.12
#>  4      65.6       1.13
#>  5      65.8       1.14
#>  6      66         1.15
#>  7      66.2       1.16
#>  8      66.4       1.17
#>  9      66.6       1.18
#> 10      66.8       1.20
#> # … with 166 more rows
#> 
#> 
#> $SMA
#> $SMA[[1]]
#> # A tibble: 15 × 5
#>    ln_STV  arc_RY SMA_line residuals fitted_axis
#>     <dbl>   <dbl>    <dbl>     <dbl>       <dbl>
#>  1  0     -0.311    0.0934   -0.0934     -1.41  
#>  2  0.693 -0.142    0.858    -0.165       0.0523
#>  3  1.10  -0.0759   1.16     -0.0577      0.756 
#>  4  1.39  -0.0320   1.35      0.0315      1.24  
#>  5  1.61   0        1.50      0.110       1.61  
#>  6  1.79   0.0743   1.83     -0.0430      2.13  
#>  7  1.95   0.0540   1.74      0.203       2.19  
#>  8  2.08   0.120    2.04      0.0365      2.62  
#>  9  2.20   0.148    2.17      0.0311      2.86  
#> 10  2.30   0.0962   1.93      0.369       2.74  
#> 11  2.40   0.180    2.31      0.0864      3.21  
#> 12  2.48   0.322    2.95     -0.467       3.94  
#> 13  2.56   0.222    2.50      0.0650      3.57  
#> 14  2.64   0.222    2.50      0.139       3.64  
#> 15  2.71   0.322    2.95     -0.244       4.16  
#> 
#> 
# }