soiltestcorr: Soil Test Correlation and Calibration in R

the soiltestcorr-package is designed to assist users on the correlation analysis of crop yield and soil test value. The soiltestcorr package estimates critical soil test values using R several different approaches such as: i) the modified ALCC model originally described by Dyson and Conyers (2013) and modified by Correndo et al. (2017); ii) the graphical Cate & Nelson approach (Cate & Nelson, 1965), iii) the statistical Cate & Nelson approach (Cate & Nelson, 1965), and iv) regression models such as linear-plateau, quadratic-plateau, and others.

See also

Author

Maintainer: Adrian A. Correndo correndo@ksu.edu (ORCID)

Authors:

Adrian A. Correndo correndo@ksu.edu (ORCID)

Austin Pearce awpearc2@ncsu.edu (ORCID)

Ignacio A. Ciampitti ciampitti@ksu.edu (ORCID)