soiltestcorr: Soil Test Correlation and Calibration

A compilation of functions designed to assist users on the correlation analysis of crop yield and soil test values. Functions to estimate crop response patterns to soil nutrient availability and critical soil test values using various approaches such as: 1) the modified arcsine-log calibration curve (Correndo et al. (2017) doi:10.1071/CP16444 ); 2) the graphical Cate-Nelson quadrants analysis (Cate & Nelson (1965)), 3) the statistical Cate-Nelson quadrants analysis (Cate & Nelson (1971) doi:10.2136/sssaj1971.03615995003500040048x ), 4) the linear-plateau regression (Anderson & Nelson (1975) doi:10.2307/2529422 ), 5) the quadratic-plateau regression (Bullock & Bullock (1994) doi:10.2134/agronj1994.00021962008600010033x ), and 6) the Mitscherlich-type exponential regression (Melsted & Peck (1977) doi:10.2134/asaspecpub29.c1 ). The package development stemmed from ongoing work with the Fertilizer Recommendation Support Tool (FRST) and Feed the Future Innovation Lab for Collaborative Research on Sustainable Intensification (SIIL) projects.

See also

Author

Maintainer: Adrian A. Correndo correndo@ksu.edu (ORCID) [copyright holder]

Authors:

Adrian A. Correndo correndo@ksu.edu (ORCID)

Austin Pearce awpearc2@ncsu.edu (ORCID)

Deanna Osmond deanna_osmond@ncsu.edu (ORCID)

Ignacio A. Ciampitti ciampitti@ksu.edu (ORCID)